+++ packages/react-native-renderer/src/__tests__/ReactNativeAttributePayload-test.js
@@ -4,6 +4,7 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
+ * @jest-environment node
  */
 'use strict';
