+++ packages/react-reconciler/src/__tests__/ReactFragment-test.js
@@ -5,6 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  *
  * @emails react-core
+ * @jest-environment node
  */
 'use strict';
