+++ packages/react-rt-renderer/src/__tests__/ReactNativeRT-test.internal.js
@@ -5,6 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  *
  * @emails react-core
+ * @jest-environment node
  */
 
 'use strict';
