+++ packages/react-test-renderer/src/__tests__/ReactTestRendererTraversal-test.js
@@ -5,6 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  *
  * @emails react-core
+ * @jest-environment node
  */
 
 'use strict';
